AccessoryIds can be Long

+++ src/main/java/io/github/hapjava/server/impl/connections/PendingNotification.java
@@ -3,11 +3,11 @@ package io.github.hapjava.server.impl.connections;
 import io.github.hapjava.characteristics.EventableCharacteristic;
 
 public class PendingNotification {
-  public int aid;
+  public long aid;
   public int iid;
   public EventableCharacteristic characteristic;
 
-  public PendingNotification(int aid, int iid, EventableCharacteristic characteristic) {
+  public PendingNotification(long aid, int iid, EventableCharacteristic characteristic) {
     this.aid = aid;
     this.iid = iid;
     this.characteristic = characteristic;
